don't delete fsxml files
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@6263 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/switch_xml.cpp
@@ -1694,7 +1694,9 @@ SWITCH_DECLARE(void) switch_xml_free(switch_xml_t xml)
 	}
 
 	if (xml->free_path) {
-		unlink(xml->free_path);
+		if (!switch_stristr(xml->free_path, ".fsxml")) {
+			unlink(xml->free_path);
+		}
 		switch_safe_free(xml->free_path);
 	}
